requirements.txt
alembic==1.13.2
SQLAlchemy==2.0.35
sqlalchemy[asyncio]
asyncpg==0.29.0
python-dotenv==1.0.1
aiogram==3.20.0
psycopg2-binary==2.9.9